Prevent one test from running in browsers
